+++ src/index.ts
@@ -14,6 +14,7 @@ export * from "./core/world/World.js";
 export * from "./core/world/WorldInteractiveObject.js";
 export * from "./core/world/WorldViewBase.js";
 export * from "./core/utilities/UpdateManager.js";
+export * from "./core/utilities/FixedStepAccumulator.js";
 export * from "./core/styles/StyleManager.js";
 export * from "./core/styles/StyledHudObject.js";
 export * from "./core/styles/SpriteStyle.js";